+++ qwt/__init__.py
@@ -63,7 +63,7 @@ from qwt.symbol import QwtSymbol as QSbl  # see deprecated section
 from qwt.text import QwtText  # noqa: F401
 from qwt.toqimage import array_to_qimage as toQImage  # noqa: F401
 
-__version__ = "0.14.7"
+__version__ = "0.15.0"
 QWT_VERSION_STR = "6.1.5"