+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # PythonQwt Releases
 
-## Version 0.14.7
+## Version 0.15.0
 
 - Added support for `QwtDateTimeScaleDraw` and `QwtDateTimeScaleEngine` for datetime axis support (see `QwtDateTimeScaleDraw` and `QwtDateTimeScaleEngine` classes in the `qwt` module)
 - Improved font rendering for rotated text in `QwtPlainTextEngine.draw` method: disabled font hinting to avoid character misalignment in rotated text